// src/SmartAbp.Vue/src/api/generated/models/SmartAbp_Application_Contracts_LowCode_Dtos_CreateOrUpdateEntityDefinitionDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { SmartAbp_Application_Contracts_LowCode_Dtos_EntityFieldDto } from './SmartAbp_Application_Contracts_LowCode_Dtos_EntityFieldDto';
export type SmartAbp_Application_Contracts_LowCode_Dtos_CreateOrUpdateEntityDefinitionDto = {
    name?: string | null;
    tableName?: string | null;
    displayName?: string | null;
    description?: string | null;
    entityType?: string | null;
    baseType?: string | null;
    namespace?: string | null;
    fields?: Array<SmartAbp_Application_Contracts_LowCode_Dtos_EntityFieldDto> | null;
};