// src/SmartAbp.Vue/src/api/generated/models/SmartAbp_Application_Contracts_LowCode_Dtos_CreateOrUpdateEntityFieldDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SmartAbp_Application_Contracts_LowCode_Dtos_CreateOrUpdateEntityFieldDto = {
    entityDefinitionId?: string;
    name?: string | null;
    displayName?: string | null;
    type?: string | null;
    length?: number | null;
    isRequired?: boolean;
    isUnique?: boolean;
    isIndexed?: boolean;
    defaultValue?: string | null;
    comment?: string | null;
    order?: number;
};