// src/SmartAbp.Vue/src/api/generated/models/SmartAbp_Application_Contracts_LowCode_SchemaValidationResult.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SmartAbp_Application_Contracts_LowCode_SchemaValidationResult = {
    isValid?: boolean;
    errors?: Array<string> | null;
    warnings?: Array<string> | null;
};